# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ManuMu Studio Authentication API
  version: 1.8.5
  description: |
    Current application-owned authentication and OAuth/OIDC endpoints.
    NextAuth's generated `/api/auth/*` endpoints are intentionally excluded.
servers:
  - url: https://auth.manumustudio.com
  - url: http://localhost:3000
paths:
...
  /api/auth/verify:
    post:
      summary: Verify an email OTP and create a NextAuth session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        "200":
          description: Verification succeeded and session cookie was set
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    const: true
        "400":
          description: Invalid, expired, or exhausted OTP
        "404":
          description: Verified user row could not be loaded
        "429":
          description: Rate limit exceeded
...
components:
  securitySchemes:
    clientBasic:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...